suppress unresolvable psalm errors
There are some problems with psalm rules and PHPUnit mocking methods.
Once I create a mock in typed attribute and trying to define mock
methods and their behavior than psalm is throwing errors like
PossiblyUndefinedMethod and MixedMethodCall. The only solution that
I found is to suppress those errors in a few cases.